놀이기구 통계 테마별 분기별 이용객 현황 조회 했긴했네 하하하 수고한 은진이,지원이

+++ hannolAdmin/src/main/java/com/sp/rides/main/RidesMain.java
@@ -3,7 +3,7 @@ package com.sp.rides.main;
 public class RidesMain {
 	private int themaCode;
 	private String themaName;
-	private long ridesCode;
+	private int ridesCode;
 	private String ridesName;
 	private int totalCount;
 
@@ -11,77 +11,60 @@ public class RidesMain {
 	private int quarter;
 	private int month;
 	private int day;
-
 	public int getThemaCode() {
 		return themaCode;
 	}
-
 	public void setThemaCode(int themaCode) {
 		this.themaCode = themaCode;
 	}
-
 	public String getThemaName() {
 		return themaName;
 	}
-
 	public void setThemaName(String themaName) {
 		this.themaName = themaName;
 	}
-
-	public long getRidesCode() {
+	public int getRidesCode() {
 		return ridesCode;
 	}
-
-	public void setRidesCode(long ridesCode) {
+	public void setRidesCode(int ridesCode) {
 		this.ridesCode = ridesCode;
 	}
-
 	public String getRidesName() {
 		return ridesName;
 	}
-
 	public void setRidesName(String ridesName) {
 		this.ridesName = ridesName;
 	}
-
 	public int getTotalCount() {
 		return totalCount;
 	}
-
 	public void setTotalCount(int totalCount) {
 		this.totalCount = totalCount;
 	}
-
 	public String getYear() {
 		return year;
 	}
-
 	public void setYear(String year) {
 		this.year = year;
 	}
-
 	public int getQuarter() {
 		return quarter;
 	}
-
 	public void setQuarter(int quarter) {
 		this.quarter = quarter;
 	}
-
 	public int getMonth() {
 		return month;
 	}
-
 	public void setMonth(int month) {
 		this.month = month;
 	}
-
 	public int getDay() {
 		return day;
 	}
-
 	public void setDay(int day) {
 		this.day = day;
 	}
+
 
 }
